requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
array-record==0.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backports.zoneinfo==0.2.1
base58==2.1.1
bigbench @ git+https://github.com/google/BIG-bench.git@761845c22056c885429efd2cfcec345ae00c1de7
black==21.12b0
BLEURT @ file://localhost//tmp/pip-req-build-0sbj_8la/bleurt/bleurt-b610120347ef22b494b6d69b4316e303f5932516.zip
blinker==1.6.2
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.3.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
chex==0.1.7
click==7.1.2
clu==0.0.9
cmake==3.26.3
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
editdistance==0.6.2
etils==1.2.0
exceptiongroup==1.1.1
faiss-cpu==1.7.4
filelock==3.12.0
flake8==6.0.0
flatbuffers==23.5.8
flax==0.6.10
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
googletrans==4.0.0rc1
grpcio==1.54.0
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.14.1
hyperframe==5.2.0
idna==2.10
immutabledict==2.2.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflate64==0.3.1
iniconfig==2.0.0
isort==5.8.0
jax==0.4.9
jaxlib==0.4.9
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.3
llvmlite==0.40.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mauve-text==0.3.0
mccabe==0.7.0
mdurl==0.1.2
mesh-tensorflow==0.1.21
ml-collections==0.1.1
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.2.2
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
plotly==5.14.1
pluggy==1.0.0
portalocker==2.7.0
promise==2.3
promptsource==0.2.3
protobuf==3.20.1
psutil==5.9.5
py7zr==0.20.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycodestyle==2.10.0
pycryptodomex==3.18.0
pydeck==0.8.1b0
pyflakes==3.0.1
pyglove==0.3.0
Pygments==2.15.1
pynndescent==0.5.10
pyparsing==3.0.9
pyppmd==1.0.0
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzstd==0.15.7
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
responses==0.18.0
RestrictedPython==6.0
rfc3986==1.5.0
rich==13.3.5
rouge-score==0.1.2
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
seqio==0.0.16
seqio-nightly==0.0.15.dev20230510
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
streamlit==0.82.0
sympy==1.12
t5==0.9.4
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorflow-text==2.12.1
tensorstore==0.1.36
termcolor==2.3.0
texttable==1.6.7
tf-slim==1.1.0
tfds-nightly==4.9.2.dev202305100044
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
tomli==1.2.3
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.29.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
umap-learn==0.5.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.4
wrapt==1.14.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
